requirements.txt
# Core Dependencies
torch==2.1.2
torchvision==0.16.2
Pillow==10.1.0

# Vision Models
groq==0.4.2
open-clip-torch==2.23.0

# Web Interface
streamlit==1.28.1
streamlit-option-menu==0.3.6

# Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2

# API & Network
requests==2.31.0
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
pydantic==2.5.1
loguru==0.7.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Development (Optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.7.1
